+++ CHANGES.rst
@@ -40,6 +40,8 @@ New Features in 23.0
   display device, ``fb-headless`` will create a headless framebuffer device
   for software rendering, and ``egl-headless`` will create a headless GPU
   device for accelerated rendering (but requires host support)
+- The `QEMUDriver` now support a ``nic`` property that can be used to create a
+  network interface when booting.
 
 Bug fixes in 23.0
 ~~~~~~~~~~~~~~~~~